--- CloudHarvestAgent/__init__.py
@@ -1,1 +1,0 @@
-from agent import *
