--- CloudHarvestAgent/agent/__init__.py
@@ -1,2 +1,0 @@
-from CloudHarvestAgent.agent.api import Api
-from CloudHarvestAgent.agent.queue import JobQueue, JobQueueStatusCodes
